Fix get_descriptor_subset
Without changing the subset based on the available distributions, the number of the physical descriptors was changed in ``SmilesMolbertModel.check_physchem_output_size`` to the number of descriptors that also appear in the distribution file. Upon creating the featurizer the next time, not all descriptors are loaded, but rather only the first k, where k is the number of descriptors that appear in the full subset and the distributions file.
Then these descriptors are again compared with the distribution file, and other descriptors may be removed as well. This happens if in the
last N-k descriptors are descriptors which appear in the distributions file.